# perf(react-router): reduce instantiations for relative routing after language service performance